+++ src/themes/int/components/legacy/TheFooter.vue
@@ -50,7 +50,7 @@
 </template>
 
 <script>
-import { mapState } from "vuex";
+import { mapGetters } from "vuex";
 
 export default {
 	data() {
@@ -58,8 +58,8 @@ export default {
 		return {};
 	},
 	computed: {
-		...mapState("auth", {
-			school: "school",
+		...mapGetters("auth", {
+			school: "getSchool",
 		}),
 		currentYear() {
 			return new Date().getFullYear();